Legacy code of home position deleted
